modules/prediction-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2
joblib==1.3.2
oracledb==2.0.0
sqlalchemy==2.0.23
geoalchemy2==0.14.2
redis==5.0.1
httpx==0.25.2
